refactor: repository layer 인터페이스 추상화
